Removing UnionProvider
