+++ provider/core/src/fork/by_key.rs
@@ -4,6 +4,7 @@
 
 //! Providers that invoke other providers based on the resource key.
 
+use crate::iter::IterableDynProvider;
 use crate::prelude::*;
 use alloc::vec::Vec;
 
@@ -266,5 +267,24 @@ where
             }
         }
         Err(DataErrorKind::MissingResourceKey.with_key(key))
+    }
+}
+
+impl<M, P> IterableDynProvider<M> for MultiForkByKeyProvider<P>
+where
+    M: DataMarker,
+    P: IterableDynProvider<M>,
+{
+    fn supported_options_for_key(
+        &self,
+        key: &ResourceKey,
+    ) -> Result<Box<dyn Iterator<Item = ResourceOptions> + '_>, DataError> {
+        for provider in self.providers.iter() {
+            let result = provider.supported_options_for_key(key);
+            if !DataError::result_is_err_missing_resource_key(&result) {
+                return result;
+            }
+        }
+        Err(DataErrorKind::MissingResourceKey.with_key(*key))
     }
 }
