+++ provider/core/src/lib.rs
@@ -143,7 +143,6 @@ mod resource;
 #[cfg(feature = "serde")]
 pub mod serde;
 pub mod struct_provider;
-pub mod union_provider;
 
 #[cfg(feature = "macros")]
 pub use icu_provider_macros::data_struct;
